refactor: socket response 분리 및 candidateList 관련 type 선언
